requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.24.0
transformers>=4.35.0
accelerate>=0.24.0
safetensors>=0.4.0
Pillow>=9.0.0
numpy>=1.23.0
tqdm>=4.64.0

# Optional – needed for SSIM-based ASR evaluation
torchmetrics>=1.0.0

# Optional – needed for LPIPS-based ASR evaluation
lpips>=0.1.4
